requirements.txt
pandas>=1.2.4
scikit-learn>=1.0.2
streamlit>=1.10.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
xgboost>=1.6.0
lightgbm>=3.3.0
imbalanced-learn>=0.8.0
python-dotenv>=0.19.0